+++ NEWS.md
@@ -1,10 +1,13 @@
 # Rgemini `develop`
 * Adds a `n_missing` function to check number of missingness.
 
+* New utility function checking user inputs
+
 # Rgemini 0.3.1
 
 * Allows the user to specify the number of digits to round percentages to in cell suppression of categorical variables, without forcing rounding to integer.
 * Allows user to render only one level for binary variables in `table1`.
+* Fixes singularity due to missing levels in `table1`.
 * Fixes `find_db_tablename` and `readmission` for DBs with foreign data wrappers.
 
 # Rgemini 0.3.0